+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Features:
 
 - When reading CSV files, assume columns in same order as defined in class if not given in file.
 - Add the ``Integer``, ``Complex``, and ``Vector`` feature types.
+- Add the ``Model.predict`` ``yield_inputs`` parameter.
 
 Bug fixes: